requirements.txt
# Flask framework for the web server
Flask

# Twilio library for WhatsApp API interaction
twilio

# Load environment variables from .env file
python-dotenv

# PostgreSQL database adapter
psycopg2-binary

# For making general HTTP requests if needed
requests


# requirements.txt (add these lines or update existing ones)
Flask
twilio
python-dotenv
psycopg2-binary
requests
Flask-SQLAlchemy # Add this for ORM integration with Flask
SQLAlchemy # Flask-SQLAlchemy might install this, but good to be explicit
requests # For making HTTP requests if needed


# For Dialogflow integration, you might need the following package
google-cloud-dialogflow